feature(form-cleanup): removed old form lib, removed old form component folder, renamed some folders for consistency
