+++ packages/lumina-core/package.json
@@ -55,7 +55,6 @@
   },
   "dependencies": {
     "classnames": "^2.5.1",
-    "react-form-component": "^2.6.6",
     "react-hook-form": "^7.52.2",
     "react-modal": "^3.16.1",
     "react-select": "^5.8.0",
