--- packages/lumina-core/src/components/form-components-RHF/input/styles.scss
@@ -1,20 +1,0 @@
-.input-container {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  &__text {
-    @include lum-typo-input; //typography
-
-    width: 100%;
-    padding: 10px 20px;
-    color: var(--lum-color-black);
-    border: 1px solid var(--lum-color-grey);
-    border-radius: 5px;
-
-    &::placeholder {
-      color: var(--lum-color-fg-weaker);
-    }
-  }
-}
