--- packages/lumina-core/src/components/form-components-RHF/slider/styles.scss
@@ -1,61 +1,0 @@
-.slider_container {
-  &__value {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-  }
-}
-
-input[type='range'] {
-
-  // Custom CSS according to Figma
-  width: 100%;
-  overflow: hidden;
-  // Removes the default appearance
-  appearance: none;
-  cursor: pointer;
-  border-radius: 16px;
-  outline: none;
-
-  &::-webkit-slider-runnable-track {
-
-    // Track - Webkit browsers
-    height: 15px;
-    background: var(--lum-color-grey);
-    border-radius: 16px;
-  }
-
-  &::-moz-range-track {
-
-    // Track - Mozilla Firefox
-    height: 15px;
-    background: var(--lum-color-grey);
-    border-radius: 16px;
-  }
-
-  &::-webkit-slider-thumb {
-
-    // Thumb - Webkit browsers
-    width: 15px;
-
-    // Custom CSS according to Figma
-    height: 15px;
-    //  Removes the default appearance
-    appearance: none;
-    background-color: var(--lum-color-white);
-    border: 2px solid var(--lum-color-accent-default);
-    border-radius: 50%;
-    box-shadow: -407px 0 0 400px var(--lum-color-accent-default);
-  }
-
-  &::-moz-range-thumb {
-
-    // Thumb - Mozilla Firefox
-    width: 15px;
-    height: 15px;
-    background-color: var(--lum-color-white);
-    border: 1px solid var(--lum-color-accent-default);
-    border-radius: 50%;
-    box-shadow: -407px 0 0 400px var(--lum-color-accent-default);
-  }
-}
