+++ packages/lumina-core/src/components/form-components/asset.png
