+++ packages/lumina-core/src/components/form-components/checkbox/styles.scss
@@ -1,34 +1,45 @@
-// The .lum__checkbox class should be added to classNames on the checkbox you want to render
-.lum__checkbox {
+.checkbox_container {
+  display: flex;
+  flex-direction: column;
   width: 100%;
 
-  .rfc-label {
-    margin-bottom: 10px;
-    font-size: 16px;
-    font-weight: 600;
-    color: var(--lum-color-black);
-  }
+  &__wrapper {
+    display: flex;
+    align-items: center;
 
-  input::before {
-    margin: -1px;
-    color: var(--lum-color-accent-default);
-    background-color: var(--lum-color-surface-default);
-    border: 1px solid var(--lum-color-accent-default);
-    border-radius: 3px;
-  }
+    &__check {
+      position: relative;
+      width: 24px;
+      height: 24px;
+      margin-bottom: 4px;
+      appearance: none;
+      cursor: pointer;
+      background-color: var(--lum-color-surface-default);
+      border: 1px solid var(--lum-color-accent-default);
+      border-radius: 3px;
 
-  input:checked::after {
-    margin: -2px;
-    color: var(--lum-color-accent-default);
-    background-color: var(--lum-color-surface-default);
-    border: 1px solid var(--lum-color-accent-default);
-    border-radius: 3px;
-  }
+      &::after {
+        position: absolute;
+        top: 40%;
+        left: 50%;
+        display: none;
+        width: 6px;
+        height: 12px;
+        content: '';
+        border: solid var(--lum-color-accent-default);
+        border-width: 0 2px 2px 0;
+        transform: translate(-50%, -50%) rotate(45deg);
+      }
 
-  span {
-    margin-top: 10px;
-    font-size: 16px;
-    font-weight: 400;
-    color: var(--lum-color-grey);
+      &:checked::after {
+        display: block;
+      }
+    }
+
+    &__text {
+      margin-left: 8px;
+      font-size: 16px;
+      color: var(--lum-color-black);
+    }
   }
 }
