--- packages/lumina-core/src/components/form-components/index.scss
@@ -1,6 +1,0 @@
-@import './input/styles';
-@import './checkbox/styles';
-@import './radio-button/styles';
-@import './switch/styles';
-@import './textarea/styles';
-@import './slider/styles';
