+++ packages/lumina-core/src/components/form-components/switch/styles.scss
@@ -1,38 +1,71 @@
-// The .lum__switch-list class should be added to classNames on the switch-list you want to render
-.lum__switch-list {
+.switch_container {
   width: 100%;
 
-  .rfc-label {
-    margin-bottom: 10px;
-    font-size: 16px;
-    font-weight: 600;
-    color: var(--lum-color-black);
-  }
+  .switch_container-toggler {
+    display: flex;
+    gap: 4px;
+    align-items: flex-start;
 
-  .rfc-checkitem {
-    font-size: 16px;
-    font-weight: 300;
-    color: var(--lum-color-black);
-  }
+    &__input {
+      position: relative;
+      min-width: 48px;
+      min-height: 22px;
+      margin: 0;
+      appearance: none;
+      cursor: pointer;
+      user-select: none;
+      background-color: var(--lum-color-grey);
+      border: 1px solid var(--lum-color-grey);
+      border-radius: 32px;
+      transition:
+        background-color 0.3s,
+        border-color 0.3s;
 
-  input[type='checkbox'] {
-    background-color: var(--lum-color-surface-default);
-    border: 1px solid var(--lum-color-grey);
+      &:checked {
+        background-color: var(--lum-color-accent-default);
+        border-color: var(--lum-color-accent-default);
 
-    &::after {
-      background-color: var(--lum-color-accent-default);
+        &::before {
+          left: unset;
+          transform: translateX(calc(100% + 12px));
+        }
+
+        &::after {
+          right: unset;
+          left: 0.4em;
+        }
+      }
+
+      &::before {
+        position: absolute;
+        top: 2px;
+        bottom: 2px;
+        left: 2px;
+        z-index: 1;
+        width: 16px;
+        height: 16px;
+        cursor: pointer;
+        content: '';
+        background-color: var(--lum-color-white);
+        border-radius: 50%;
+        transition: all 0.3s;
+      }
+
+      &::after {
+        position: absolute;
+        top: 0.5em;
+        right: 0.4em;
+        bottom: 0.5em;
+        display: block;
+        font-size: 10px;
+        line-height: 1;
+        color: var(--lum-color-white);
+      }
+
+      &:focus {
+        outline: 2px solid var(--lum-color-accent-default);
+        outline-offset: 2px;
+      }
     }
-  }
-
-  input:checked {
-    color: var(--lum-color-surface-default);
-    border: 1px solid var(--lum-color-accent-default);
-  }
-
-  span {
-    margin-top: 10px;
-    font-size: 16px;
-    font-weight: 400;
-    color: var(--lum-color-grey);
   }
 }
