+++ packages/lumina-core/src/components/form-components/textarea/index.tsx
@@ -1,12 +1,28 @@
-import { TextArea, type ControlLogicProps } from 'react-form-component'
 import cx from 'classnames'
+import { TextareaHTMLAttributes } from 'react'
+import { LabelTitle } from '../utils/utility-components/label'
+import { HelpText } from '../utils/utility-components/help'
 
-export type { ControlLogicProps }
+interface ICheckbox extends TextareaHTMLAttributes<HTMLTextAreaElement> {
+  className?: string
+  label?: string
+  help?: string
+  labelClassName?: string;
+  textAreaClassName?: string;
+  helpClassName?: string;
+}
 
-export const LumTextArea = (props: ControlLogicProps) => {
-  const defaultClass = 'lum__text-area'
-  const className = cx(defaultClass, props.className)
-  const name = props.name
-
-  return <TextArea {...props} name={name} className={className} />
+/**
+ * @label A text that is positioned on top of the input range
+ * @help  A helper text position below the slider
+ * @returns
+ */
+export const TextArea = ({ className, label, help, labelClassName, textAreaClassName, helpClassName, ...rest }: ICheckbox) => {
+  return (
+    <div className={cx('text-area_container', className)}>
+      <LabelTitle label={label} className={labelClassName} />
+      <textarea className={cx('text-area_container__text', textAreaClassName)} {...rest} />
+      <HelpText className={helpClassName} help={help} />
+    </div>
+  )
 }
