+++ packages/lumina-core/src/components/form-components/textarea/styles.scss
@@ -1,28 +1,23 @@
-// The .lum__text-area class should be added to classNames on the text-area you want to render
-.lum__text-area {
+.text-area_container {
+  position: relative;
+  display: flex;
+  flex-direction: column;
   width: 100%;
 
-  .rfc-label {
-    margin-bottom: 10px;
-    font-size: 16px;
-    font-weight: 600;
-    color: var(--lum-color-black);
-  }
+  &__text {
+    @include lum-typo-input; // typography
 
-  textarea.rfc-input {
+    width: 100%;
+    min-height: 70px;
+    max-height: 450px;
+    padding: 10px 20px;
     color: var(--lum-color-black);
-    border-color: var(--lum-color-grey);
+    resize: vertical;
+    border: 1px solid var(--lum-color-grey);
     border-radius: 5px;
 
     &::placeholder {
       color: var(--lum-color-fg-weaker);
     }
-  }
-
-  span {
-    margin-top: 10px;
-    font-size: 16px;
-    font-weight: 400;
-    color: var(--lum-color-grey);
   }
 }
