+++ packages/lumina-core/src/components/form-components/utils/utility-components/help/styles.scss
