+++ packages/lumina-core/src/components/styles.scss
@@ -8,5 +8,4 @@
 @import './tabs/pages/styles';
 @import './tabs/treeView/styles';
 @import './title/styles';
-@import './form-components/index';
-@import './form-components-RHF/';
+@import './form-components/styles';
